Add GitHub And Google Nextauth Providers (#15)
This pull request adds GitHub and Google Nextauth providers which allows users to sign into the website using accounts from those websites via OAuth `2.0`. As of right now, there is not functionality to link accounts. Instead, if an account exists with the same email, the user will have to log into whatever provider they used with that email. Account linking will be a feature we'll likely add in the future.

Additionally, the following changes were made.

- Updated NodeJS types package to `18.17.1` since >=18.17.1 NodeJS server is required for new authentication to work.
- Updated VS Code task's dev server to listen on `0.0.0.0` properly.
- Moved session interface definition from `src/types/next-auth.d.ts` to `src/server/auth.ts`.